Enzo/reservas

+++ backend/src/main/java/com/temreserva/backend/temreserva_backend/business/RestaurantBusiness.java
@@ -173,9 +173,12 @@ public class RestaurantBusiness {
             r.setUpdateDate(LocalDateTime.now());
             r.setPhoneNumber(restaurant.getPhoneNumber() == null ? r.getPhoneNumber() : restaurant.getPhoneNumber());
             r.setPayment(restaurant.getPayment() == null ? r.getPayment() : restaurant.getPayment());
-            restaurantDateTimeRepository.findByRestaurant(r).forEach(a -> restaurantDateTimeRepository.delete(a));
-            restaurant.getRestaurantDateTime().forEach(rest -> createRestaurantDateTime(r, rest));
-            credentialBusiness.updateEmailByID(r.getCredential().getId(), restaurant.getEmail());
+            if(restaurant.getRestaurantDateTime() != null) {
+                restaurantDateTimeRepository.findByRestaurant(r).forEach(a -> restaurantDateTimeRepository.delete(a));
+                restaurant.getRestaurantDateTime().forEach(rest -> createRestaurantDateTime(r, rest));
+            }            
+            if(restaurant.getEmail() != null)
+                credentialBusiness.updateEmailByID(r.getCredential().getId(), restaurant.getEmail());
             return restaurantRepository.save(r);
         }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                 Enumerators.apiExceptionCodeEnum.RESTAURANT_NOT_FOUND.getEnumValue()));
@@ -210,9 +213,11 @@ public class RestaurantBusiness {
                     .findCurrentPeopleModelByRestaurant(restaurant.getId(), LocalDate.now());
             List<RestaurantDateTimeModel> dateTime = getListRestaurantDateTimeModel(
                     restaurantDateTimeRepository.findByRestaurant(restaurant));
-            return RestaurantModel.builder().id(restaurant.getId()).email(restaurant.getCredential().getEmail())
-                    .profileImage(img).cnpj(restaurant.getCnpj()).cleaning(restaurant.getCleaning())
-                    .address(getAddressModelFromAddress(address)).restaurantName(restaurant.getRestaurantName())
+            return RestaurantModel.builder().id(restaurant.getId())
+                    .actualNumberOfPeople(restaurant.getActualNumberOfPeople())
+                    .email(restaurant.getCredential().getEmail()).profileImage(img).cnpj(restaurant.getCnpj())
+                    .cleaning(restaurant.getCleaning()).address(getAddressModelFromAddress(address))
+                    .restaurantName(restaurant.getRestaurantName())
                     .currentPeople(getListCurrentPeopleModel(currentPeople)).restaurantImages(lstImages)
                     .maxNumberOfPeople(restaurant.getMaxNumberOfPeople()).averageStars(restaurant.getAverageStars())
                     .payment(restaurant.getPayment()).phoneNumber(restaurant.getPhoneNumber())
@@ -287,7 +292,8 @@ public class RestaurantBusiness {
                         .findCurrentPeopleModelByRestaurant(restaurant.getId(), LocalDate.now());
                 List<RestaurantDateTimeModel> dateTime = getListRestaurantDateTimeModel(
                         restaurantDateTimeRepository.findByRestaurant(restaurant));
-                RestaurantModel model = RestaurantModel.builder().id(restaurant.getId())
+                RestaurantModel model = RestaurantModel.builder()
+                        .actualNumberOfPeople(restaurant.getActualNumberOfPeople()).id(restaurant.getId())
                         .email(restaurant.getCredential().getEmail()).profileImage(img).cnpj(restaurant.getCnpj())
                         .cleaning(restaurant.getCleaning()).address(getAddressModelFromAddress(address))
                         .restaurantName(restaurant.getRestaurantName())
